// src/modules/auth/auth.schema.ts
import { z } from "zod";

export const registerSchema = z.object({
  name: z.string().min(2),
  email: z.string().email(),
  password: z.string().min(6),
  role: z.enum(["PM", "INTERNAL", "CLIENT_GUEST"]),
  department: z.enum(["PRODUCT", "UI_UX", "FRONTEND", "BACKEND"]).optional(),
});

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string(),
});